// ui/src/client/types.gen.ts
export type FilePresign = {
    /**
     * File Name
     */
    file_name: string;
    /**
     * Mime Type
     */
    mime_type: string;
    role: FileRole;
};
// ...
export const FileRole = {
    TEMPLATE: 'template',
    PREVIEW_PDF: 'preview_pdf',
    ARTEFACT: 'artefact',
    INPUT: 'input'
} as const;

/**
 * FileRole
 */
export type FileRole = typeof FileRole[keyof typeof FileRole];